Adds Windows support
* Updates CMakeLists files to handle MSVC options, but preserve
  all-warnings-are-errors behaviour.
* Fixes C++ reported as errors by MSVC
+ Adds Visual Studio IDE files to .gitignore

+++ src/proto/inc/proto/delimited_protobuf_stream.hh
@@ -53,7 +53,7 @@ std::istream& operator>>(std::istream& is, DelimitedProtobufStream<T>& stream)
         }
 
         is.read((char*)stream.buffer, size);
-        stream.valid = stream.message.ParseFromArray(stream.buffer, size);
+        stream.valid = stream.message.ParseFromArray(stream.buffer, (int)size);
     } else {
         stream.valid = false;
     }
